update to lua 5.4.7

+++ scripts/update_lua.sh
@@ -1,6 +1,10 @@
 #! /bin/sh
 
-VER=5.4.6
+# stop on error
+set -e
+set -x
+
+VER=5.4.7
 
 SRC=lua-$VER
 DEST=../libs/lua
@@ -10,7 +14,7 @@ DEST=../libs/lua
 cd $(dirname $0)
 
 # get latest source
-curl -O http://www.lua.org/ftp/$SRC.tar.gz
+curl -O https://www.lua.org/ftp/$SRC.tar.gz
 tar -xvf $SRC.tar.gz
 
 # remove lua standalone console and compiler sources
@@ -20,7 +24,7 @@ rm -v $SRC/src/lua.c $SRC/src/luac.c
 mkdir -p $DEST
 
 # copy license
-cp -v $SRC/COPYRIGHT $DEST
+test -f $SRC/COPYRIGHT && cp -v $SRC/COPYRIGHT $DEST
 
 # copy sources
 cp -v $SRC/src/*.h $DEST
